made narbar into an aside navbar
